SRCH-3538: added test for Header and using const and func for header and footer data
