+++ app/javascript/components/Footer/Footer.css
@@ -1,0 +1,4 @@
+/* Hiding the seconday section as its not needed - SRCH-3539 */
+#serp-footer-wrapper .usa-footer__secondary-section{
+  display: none;
+}
