+++ app/javascript/components/Identifier/Identifier.tsx
@@ -1,7 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Identifier as UswdsIdentifier, IdentifierMasthead, IdentifierLinks, IdentifierLogos, IdentifierLogo, IdentifierIdentity, Link, IdentifierGov, IdentifierLinkItem, IdentifierLink } from '@trussworks/react-uswds';
 
-//import circleSvg from 'uswds/src/img/circle-gray-20.svg';
+//this is just a dummy logo for UI purposes - to be dynamic
+const logoImg = "https://search.gov/assets/gsa-logo-893b811a49f74b06b2bddbd1cde232d2922349c8c8c6aad1d88594f3e8fe42bd097e980c57c5e28eff4d3a9256adb4fcd88bf73a5112833b2efe2e56791aad9d.svg";
+
 interface IdentifierProps {
 }
 
@@ -25,7 +27,7 @@ export const Identifier = (props: IdentifierProps) => {
             <IdentifierLogo href="">
               <img
                 className="usa-identifier__logo-img"
-                src="https://search.gov/assets/gsa-logo-893b811a49f74b06b2bddbd1cde232d2922349c8c8c6aad1d88594f3e8fe42bd097e980c57c5e28eff4d3a9256adb4fcd88bf73a5112833b2efe2e56791aad9d.svg"
+                src={logoImg}
                 alt="<Parent agency> logo"
               />
             </IdentifierLogo>
